feat(Renderer): Hoàn thiện Shadow Mapping và tích hợp FXAA
Bản cập nhật lớn này giới thiệu hai tính năng quan trọng: đổ bóng động (Dynamic Shadow Mapping) và khử răng cưa FXAA, đồng thời tái cấu trúc và dọn dẹp mã nguồn để tăng tính dễ bảo trì.

**Tính năng chính:**

1.  **Dynamic Shadow Mapping:**
    *   Thêm `ShadowMapPass` mới để render shadow map từ góc nhìn của nguồn sáng.
    *   Cập nhật `LightingPass` để lấy mẫu shadow map, tính toán và áp dụng bóng đổ vào kết quả chiếu sáng cuối cùng, tạo ra các vùng bóng đổ chân thực.
    *   Mở rộng `Light` và `LightManager` để quản lý `lightSpaceMatrix` và tài nguyên shadow map.

2.  **Khử răng cưa FXAA & Hậu kỳ:**
    *   Thay thế `CompositePass` cũ bằng một shader mới tích hợp thuật toán khử răng cưa FXAA.
    *   Thêm bước làm nét ảnh (Contrast Adaptive Sharpening) để phục hồi chi tiết bị mờ do FXAA.
    *   Shader hậu kỳ giờ đây kết hợp (Scene + Bloom + ToneMap + FXAA) trong một lần xử lý duy nhất.

**Tái cấu trúc và Cải tiến:**

*   **Đổi tên:** Các shader và biến liên quan đến G-Buffer được đổi tên từ "RTT" thành "Geometry" cho rõ nghĩa.
*   **Ánh sáng:** Cấu hình lại hệ thống ánh sáng mặc định thành dạng Key-Fill/Rim để có chất lượng chiếu sáng tốt hơn.
*   **Dọn dẹp:** Thêm và cải thiện comment tiếng Việt, loại bỏ các "magic number".

+++ VulkanLearn2/Scene/MaterialManager.cpp
@@ -151,7 +151,7 @@ uint32_t MaterialManager::LoadMaterial(const MaterialRawData& materialRawData)
 	// Thêm vật liệu đã cấu hình vào danh sách tất cả các vật liệu được quản lý.
 	m_Handles.allMaterials.push_back(material);
 	// Trả về chỉ số của vật liệu vừa được thêm vào.
-	return m_Handles.allMaterials.size() - 1;
+	return static_cast<uint32_t>(m_Handles.allMaterials.size() - 1);
 }
 
 VulkanDescriptor* MaterialManager::GetDescriptor()
